requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
appenginetaskutils==0.4.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.5
Django==4.1.2
django-debug-toolbar==3.7.0
django-ipware==4.0.2
executing==1.1.1
frozenlist==1.3.1
geoip2==4.6.0
google-api-core==2.10.2
google-auth==2.12.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
GoogleAppEngineCloudStorageClient==1.9.22.1
icecream==2.1.3
idna==3.4
maxminddb==2.2.0
multidict==6.0.2
postgres==4.0
protobuf==4.21.7
psycopg2-binary==2.9.4
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygeoip==0.3.2
Pygments==2.13.0
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
Werkzeug==0.12.2
yarl==1.8.1
